fix_home_work
